Split stargazer and radar shells into controller hooks.
Extract useStargazerController + StargazerCommandCenter and useRadarController so page/shell files stay presentational, matching the space-weather ownership pattern.
